+++ packages/core/src/index.ts
@@ -11,6 +11,7 @@ export * from "./format.js";
 export * from "./analytics.js";
 export * from "./guestbook.js";
 export * from "./presence.js";
+export * from "./music.js";
 export * from "./assets.js";
 export * from "./resolve.js";
 export * from "./limits.js";